feat: add api request and query for fetching docs
