+++ src/constants/TanStackQueryKeys.ts
@@ -1,7 +1,6 @@
 export const QK_COMPETENCIES = 'competencies';
 export const QK_PRODUCTS = 'products';
 export const QK_PROJECTS = 'projects';
-
+export const QK_DOCS = 'docs';
 export const QK_USER_PROFILE = 'userProfile';
-
 export const QK_REG = 'registration';
