+++ src/types/forms.ts
@@ -1,0 +1,15 @@
+// forms data
+export type RegFormData = {
+  email: string;
+  password: string;
+};
+
+export type LoginFormData = {
+  email: string;
+  password: string;
+};
+
+export type TChangedPwdData = {
+  current_password: string;
+  new_password: string;
+};
